+++ src/api/dart/mod.rs
@@ -599,7 +599,7 @@ impl TryFrom<DartValueArg<Self>> for Option<bool> {
 
 /// Error of converting a [`DartValue`] to the concrete type.
 #[derive(Debug, Display)]
-#[display(fmt = "expected `{}`, but got: `{:?}`", expectation, value)]
+#[display("expected `{expectation}`, but got: `{value:?}`")]
 pub struct DartValueCastError {
     /// Expected type description. Like a [`String`] or an `Option<i64>`.
     expectation: &'static str,